whisper-ai/requirements.txt
# Whisper AI - Comprehensive Requirements
# All dependencies updated for Python 3.12 compatibility

# ===== Core Framework =====
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0

# ===== Deep Learning & AI =====
# Updated for Python 3.12 compatibility
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
diffusers==0.25.0
transformers==4.37.0
accelerate==0.26.0
sentencepiece==0.1.99

# ===== Data Processing =====
# Updated for Python 3.12 - versions with prebuilt wheels
numpy==1.26.0
pandas==2.1.4

# ===== HTTP & Networking =====
# Updated for Python 3.12 compatibility
httpx>=0.27.0
aiohttp==3.9.1
requests==2.31.0

# ===== Web Scraping & Parsing =====
beautifulsoup4==4.12.3
lxml==5.1.0

# ===== Logging =====
loguru==0.7.2

# ===== Security & Encryption =====
cryptography==42.0.0

# ===== Data Formats =====
pyyaml==6.0.1

# ===== Streaming Support =====
sse-starlette==2.0.0

# ===== Image Processing =====
Pillow==10.2.0
opencv-python==4.9.0.80

# ===== AI Service APIs =====
google-genai==1.56.0
google-auth==2.45.0
openai==1.3.9

# ===== Text-to-Speech =====
gTTS==2.5.0
pyttsx3==2.90

# ===== Utilities =====
python-multipart==0.0.6
python-dotenv==1.0.0
psutil==5.9.6  # For system analytics

# ===== Database (optional, for persistence) =====
faiss-cpu==1.8.0  # Minimum version for Python 3.12
libsql==0.1.11

# ===== Redis (optional, for token session management) =====
redis==5.0.1

# ===== Embeddings (optional, for knowledge base) =====
sentence-transformers==2.3.1

# ===== HuggingFace Dataset Sync =====
datasets==2.16.1
huggingface_hub==0.20.2
